Find rankings of journals from the Monash Business School, ABDC, CORE, SCImago or ERA2010.

Given a list of journal titles, this function will return their ranking from various lists. Data sets used are:

monash:

Monash Business School

abdc:

Australian Business Deans' Council

era2010:

ERA 2010

core:

CORE

scimago:

SCImago

This function is used in the Journal rankings shiny app.

Usage

journal_ranking(
  title,
  source = c("monash", "abdc", "era2010", "core", "scimago"),
  fuzzy = TRUE,
  only_best = length(title) > 1,
  ...
)

Arguments

title

A character vector containing (partial) journal names.

source

A character string indicating which ranking data base to use. Default "monash".

fuzzy

Should fuzzy matching be used. If FALSE, partial exact matching is used. Otherwise, full fuzzy matching is used.

only_best

If TRUE, only the best matching journal is returned.

...

Other arguments are passed to agrepl (if fuzzy is TRUE), or grepl otherwise.

Value

A data frame containing the journal title, rank and source for each matching journal.

Author

Rob J Hyndman

Examples

# Return ranking for individual journals or conferences
journal_ranking("Annals of Statistics")
#> # A tibble: 1 × 2
#>   title                rank    
#>   <chr>                <fct>   
#> 1 Annals of Statistics Group 1+
journal_ranking("Annals of Statistics", "abdc")
#> # A tibble: 1 × 2
#>   title                rank 
#>   <chr>                <ord>
#> 1 Annals of Statistics A*   
journal_ranking("International Conference on Machine Learning")
#> # A tibble: 1 × 2
#>   title                                        rank   
#>   <chr>                                        <fct>  
#> 1 International Conference on Machine Learning Group 1
journal_ranking("International Conference on Machine Learning", "core")
#> # A tibble: 3 × 2
#>   title                                                         rank 
#>   <chr>                                                         <ord>
#> 1 International Conference on Machine Learning                  A*   
#> 2 International Conference on Machine Learning and Cybernetics  NA   
#> 3 International Conference on Machine Learning and Applications C    
journal_ranking("R Journal", "scimago", only_best = TRUE)
#> # A tibble: 1 × 2
#>   title     rank 
#>   <chr>     <chr>
#> 1 R Journal Q1